pyproject.toml
[project]
name = "rag-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.52.0",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
